+++ sources/novelarchive/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog - Novel Archive
 
+## [0.4] - 2026-09-23
+- Added novel listing
+- Added genre listing
+
 ## [0.3] - 2026-09-21
 - Optimised for Fast Searching