+++ sources/novelbins/CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog - Novel Bins
 
+## [0.3] - 2026-09-23
+- Fixed novel listings
+- Fixed genre listings
+
 ## [0.2] - 2026-09-21
 - Initial release.