+++ sources/novelfire/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog - Novel Fire
 
+## [0.4] - 2026-09-23
+- Updated to match template
+
 ## [0.3] - 2026-09-21
 - Initial release.