+++ sources/novelphoenix/CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog - Novel Phoenix
 
+## [0.3] - 2026-09-23
+- Updated to match novelfire template
+
 ## [0.2] - 2026-09-21
 - Initial release.